refactor: extract lit-billing-core from lit-api-server stripe.rs
Move Stripe HTTP client, customer-identity primitives, balance read,
and reporting helpers into a new lit-billing-core crate. lit-payments
(separate non-TEE service, coming in a follow-up PR) will depend on
the same crate so the customer-identity invariant
(metadata.wallet_address keys the Stripe customer) is shared.

What moves to lit-billing-core:
  - StripeClient (creds + reqwest, no caching)
  - parse_stripe_response + StripeResponse
  - customer::find_or_create_by_wallet, set_email
  - balance::fetch
  - reporting::list_all_customers / list_balance_transactions_since /
    aggregate_report_rows / Report{Customer,BalanceTx,Row}
  - format::cents_to_display / unix_to_utc_date

What stays in lit-api-server/src/stripe.rs:
  - StripeState (caches + publishable key)
  - resolve_wallet_address (depends on on-chain accounts module)
  - Cached wrappers: get_customer_by_wallet, get_credit_balance
  - Charge flows: charge / charge_management / charge_lit_action_time
  - PaymentIntent flow: create_payment_intent / confirm_payment_and_credit
  - register_customer_email, invalidate_wallet_cache
  - Reporting wrappers that take &StripeState

Pure refactor: no behavior change. lit-api-server's public stripe API
is unchanged; callers in lit_api_server::stripe::* still compile and
pass tests as before. cargo check + cargo test --lib (211 tests) +
cargo test --bin stripe_report (14 tests) + clippy -D warnings all
clean.
