+++ lit-api-server/Cargo.toml
@@ -35,6 +35,7 @@ sha3 = "0.10"
 ipfs-hasher = "0.13.0"
 elliptic-curve = "0.13.8"
 k256 = { version = "0.13.4", features = ["ecdsa"] }
+lit-billing-core = { path = "../lit-billing-core" }
 lit-core = { path = "../lit-core/lit-core" }
 lit-core-derive = { path = "../lit-core/lit-core-derive" }
 lit-observability = { path = "../lit-core/lit-observability", features = ["channels"] }
