+++ lit-billing-core/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "lit-billing-core"
+version = "0.1.0"
+edition = "2024"
+description = "Shared Stripe client + customer/balance primitives for Lit billing services"
+license = "Apache-2.0"
+
+[dependencies]
+anyhow = "1.0"
+reqwest = { version = "0.12.10", features = ["json"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+tracing = "0.1"
